add fix for CIFS client memory leak
As reported in the community forum [0], there currently is a memory
leak in the CIFS client code. Reproduced by running a backup with CIFS
target storage:

> while true; do vzdump 101 --storage cifs --prune-backups keep-last=1; echo 3 > /proc/sys/vm/drop_caches; done

A fix was found on the kernel mailing list tagged for stable v6.6+
and it does solve the issue, but is not yet included in any (stable)
kernels.

[0]: https://forum.proxmox.com/threads/147603/post-682388

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>