fix #5683: netfs: reset subreq iov iter before tail clean
Fixes rare read corruption issues using the in kernel ceph client.

On incomplete read requests, the clean tail flag should make sure to
zero fill the remaining bytes for the subrequest.
If the iov iterator is not at the correct position, e.g., because the
subreq->transferred was not yet updated, this can however zero fill
downloaded data, corrupting the read content.

Link to issue:
https://bugzilla.proxmox.com/show_bug.cgi?id=5683

Link to upstream issue:
https://bugzilla.kernel.org/show_bug.cgi?id=219237

Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
 [ TL: mention an specific example for subreq misalignment ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>