patches: kernel: switch to using full index for patch files
git will automatically change the length of the index hashes when
formatting a patch depending on what references are present in the
submodule. After pulling in the stable tags today, git wanted to add
a character to all hashes for me. Use --full-index when generating the
patches to avoid such issues in the future.

No functional change intended.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>