buildsys: add explicit DEB_VERSION variable in Makefile
Can be used to generate a package for testing a particular bugfix, or
new version, which can be shared with a small audience, without
reusing the current package version, or bumping to a version that
looks like our released kernels[0].

previously this could be accomplished by setting the PKGREL variable
to the revision and temporary suffix. PKGREL got dropped in:
de1853080134 ("bump version to 6.2.16-8")
as each new version we release needs to have a ABI bump anyways (due
to module signing).

[0] e.g. new versions of submodules (ZFS), or packages shared in the
temp-dir of our CDN nodes

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>