backport fix for network crashes when exiting ovs-tcpdump
A user in enterprise support reported a low (<1/50) chance of network
crash (with soft lockups reported) when exiting ovs-tcpdump (a tcpdump
wrapper provided by Open vSwitch) which they could only resolve by
rebooting the host.

After reporting the issue upstream with a reproducer [1], an OVS
developer submitted a kernel patch which is now included 6.13 and some
stable kernels. With this patch, the reproducer does not seem to
trigger the issue anymore. Hence, backport the patch.

[1] https://mail.openvswitch.org/pipermail/ovs-discuss/2025-January/053423.html

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>