update source and patches to Ubuntu 25.04 Plucky based 6.14 kernel
Provide a newer – currently – opt-in kernel that will form the base
for Ubuntu's 25.04 release.

Using the current relevant tag Ubuntu-6.14.0-12.12 (the newer 13.13
tag includes only some packaging changes that do not affect us) from
the upstream repo [0].

Update the patches, drop those that got applied upstream or are simply
not relevant for the newer kernel anymore. Update the ZFS submodule to
include a cherry-pick of the two Linux 6.14 compat patches that
already got applied to ZFS upstream [1].

[0]: https://git.launchpad.net/~ubuntu-kernel/ubuntu/+source/linux/+git/plucky/
[1]: https://github.com/openzfs/zfs/pull/17026

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>