cherry-pick additional fix for AMD Epyc genua systems
reports in our community forum say, that some systems still run into a
boot-loop with the latest proxmox-kernel-6.14 (6.14.0-2):
https://forum.proxmox.com/threads/164497/post-762978
Pulling in this patch here (or preventing the module from being loaded) fixes
the issue for them.

As our test-system boots fine with this patch and without it (based on
my tests for including the 2 patches, that are also cherry-picked for
kernel.org stable 6.14.2) I think adding this additionally should not
pose much potential for regression.

Fixes: 4a6063d2f956 ("cherry-pick fixes for AMD Epyc genua systems")
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20250414145705.1262110-1-s.ivanov@proxmox.com