make: use grouped target for DEB's and execute lintian only for deb target
A group target is avoiding odd behavior due to make not being in the
picture about what actually generates which DEB target and we often
move lintian out to the generic 'all' or 'deb' target over the
specific ones for better separation. We might want to make a lintian
specific target that the upload one could depend on, but in practice
this should not matter as no maintainer should do a straight "build to
upload" action.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>