d/rules: fwcheck: use C.utf8 locale for sort output
This is mostly to avoid line re-ordering noise in the fwlist-previous
file checked in to git, e.g. when multiple environment (from one or
more devs) are used to build the kernel and check the changes into
git. This does NOT have any effect on the actual check, as we always
sorted both inputs with the same sort call on every execution of the
check, so the sorting was always normalized to the locale of the build
system and thus the check never resulted in false positives.
Checking in changes could lead to more noise in the diff though and
thus potential false impression for what actually changed firmware
wise.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>